Add @mark to Player, make game logic anonymous
I refactored the game logic to work with a turn switch instead of
fixed roles (i.e. player1, player2). To implement that a @mark was
needed to #mark the board. I did it so the game would be able to stop
whenever one of the players wins. The player is now initiated with a
mark in greetings().